devices: uevent - allow internal subscription to SEQNUM
Callbacks can be registered to be triggered when uevents are
guaranteed to be in sync with state read from /sys. For now the
only user is a callback to be called when uevents are up-to-date
with the enumeration of /sys. We call this event 'settle'. For
now the callback is a noop.

Signed-off-by: Tom Gundersen <teg@jklm.no>