devcies: device - track devices
All kernel devices can be discovered in /sys, and they are updated
through uevents. To avoid any races, we have to first acquire a
consistent state and then update it when new uevents come in.

First subscribe to uevents, then obtain a preliminary state by
enumerating /sys, then read out the uevent seqnum that
correpsonds to the last change to /sys. Finally, update the
preliminary state by processing uevents up-to and including
the noted seqnum. At this point the state is guaranteed to
be consistent, and it can start to be used.

Signed-off-by: Tom Gundersen <teg@jklm.no>