devices: device - verify subsystem
Before adding a new subsystem, verify that it is correct. We do
receive uevents with invalid SUBSYSTEM strings, which we would
never encounter through enumerating /sys. To avoid confusion,
make sure we treat these events as invalid and never forward
them, or any other events pertaining to the same device.

It could be that a valid device is quickly added and removed again
in which case we would suppress all events about the device. We are
not making any promises about delivering events about devices that
are already gone by the time the event arrives, so that is fine.

Internally we still track the devices, but mark them as invalid. This
avoids us re-resolving the invalid subsystem more than once perdevice.

Signed-off-by: Tom Gundersen <teg@jklm.no>